Add Linux no-oping
This is useful if you want to use these rules on swift_binary or
cc_binary targets, this way you don't have to manually exclude them at
the `deps` site. I'm not sure in general if this is a bit too weird, but
given that we're vending pre-compiled binaries, and the alternative is
force the users to add this `select()` instead, we can try this out.